lwcolor: Add 'void' to the definition of func_init().
According to ISO/IEC 9899:1999 6.7.5.3 Function declarators (including prototypes),
An empty list in a function declarator that is part of a definition of that function specifies that the function has no parameters.
The empty list in a function declarator that is not part of a definition of that function specifies that no information about the number or types of the parameters is supplied.

And, 6.11.6 Function declarators,
The use of function declarators with empty parentheses (not prototype-format parameter type declarators) is an obsolescent feature.

Therefore, in the range of C99, void func(){...} is equal to void func(void){...} while void func(); is not equal to void func(void);.
However, thinking about the future, this change is preferable.